#136 - move calculate_fu to yaku_calculator
