+++ mahjong/yaku_types.py
@@ -1,5 +1,4 @@
 import copy
-import math
 from typing import List, Optional, Callable
 from collections import defaultdict
 from itertools import combinations
@@ -60,103 +59,6 @@ class YakuTypes(ABC):
     @abstractmethod
     def get_all_evals(self):
         return NotImplemented
-
-    def calculate_fu(self):
-
-        if self.player.menzenchin and self.is_ron:
-            fu = 30
-        else:
-            fu = 20
-
-        if 'pinfu' in self.total_yaku:
-            if self.total_han == 1:
-                fu = 30  # avoid 1 han 20 fu
-            return fu
-        elif 'chiitoitsu' in self.total_yaku:
-            fu = 25
-            return fu
-
-        honor_tiles, terminal_tiles = Tile.get_yaochuuhai()
-        yaochuuhai = honor_tiles + terminal_tiles
-        for huro in self.player.kabe:
-            if huro.naki_type == Naki.PON:
-                if huro.tiles[0] in yaochuuhai:
-                    fu += 4
-                else:
-                    fu += 2
-            elif huro.naki_type == Naki.ANKAN:
-                if huro.tiles[0] in yaochuuhai:
-                    fu += 32
-                else:
-                    fu += 16
-            elif huro.naki_type in [Naki.DAMINKAN, Naki.CHAKAN]:
-                if huro.tiles[0] in yaochuuhai:
-                    fu += 16
-                else:
-                    fu += 8
-
-        def calc_wait_pattern_fu(ankous: List[Tile],
-                                 shuntsus: List[Tile],
-                                 jantou: Tile) -> int:
-            wait_pattern_fu = 0
-
-            # 暗刻
-            for tile in ankous:
-                if tile in yaochuuhai:
-                    wait_pattern_fu += 8
-                else:
-                    wait_pattern_fu += 4
-
-            # 雀頭
-            if jantou.suit == 0:
-                if jantou.rank == self.bakaze.value:
-                    wait_pattern_fu += 2
-                if jantou.rank == self.player.jikaze.value:
-                    wait_pattern_fu += 2
-
-            # 聽牌形式
-            def tenpai_add_fu() -> bool:
-                if self.player.agari_tile == jantou:  # 單騎聽
-                    return True
-                for shuntsu in shuntsus:
-                    if self.player.agari_tile == shuntsu[1]:  # 坎張聽
-                        return True
-                    elif (
-                        (self.player.agari_tile == shuntsu[0]
-                         and shuntsu[2].rank == 9)
-                        or (self.player.agari_tile == shuntsu[2]
-                            and shuntsu[0].rank == 1)
-                    ):
-                        return True  # 邊張聽
-                return False
-
-            if tenpai_add_fu():
-                wait_pattern_fu += 2
-
-            return wait_pattern_fu
-
-        # separate sets
-        tenpai_tiles = check_tenpai(self.player.hand, self.player.kabe)
-        player_huro_n = len(self.player.kabe)
-        wait_patterns = {}
-        for idx, pot_agari_tile in enumerate(tenpai_tiles):
-            tmp_agari_hand = copy.deepcopy(self.player.hand)
-            tmp_agari_hand[pot_agari_tile.index] += 1
-
-            ankous, shuntsus, jantou = separate_sets(tmp_agari_hand,
-                                                     player_huro_n)
-            wait_patterns[idx] = [ankous, shuntsus, jantou]
-
-        wait_pattern_fus = []
-        for idx in wait_patterns.keys():
-            [ankous, shuntsus, jantou] = wait_patterns[idx]
-            wait_pattern_fus.append(
-                calc_wait_pattern_fu(ankous, shuntsus, jantou))
-
-        wait_pattern_fu = max(wait_pattern_fus)
-        fu += wait_pattern_fu
-        # round up
-        return int(math.ceil(fu / 10.0)) * 10
 
     @property
     def player(self):
